+++ examples/GeocodingWidget.html
@@ -10,63 +10,59 @@
     <script src="../packages/browser/dist/RUN_MODE/bundle.js"></script>
 
     <script type="text/javascript">
-      udvizBrowser.loadMultipleJSON([
-        './assets/config/extent_lyon.json',
-        './assets/config/frame3D_planars.json',
-        './assets/config/server/geocoding_server.json',
-        './assets/config/layer/base_maps.json',
-      ]).then((configs) => {
-        // http://proj4js.org/
-        // define a projection as a string and reference it that way
-        // the definition of the projection should be in config TODO_ISSUE
-        udvizBrowser.proj4.default.defs(
-          configs['extent_lyon'].crs,
-          '+proj=lcc +lat_1=45.25 +lat_2=46.75' +
-            ' +lat_0=46 +lon_0=3 +x_0=1700000 +y_0=5200000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs'
-        );
+      udvizBrowser
+        .loadMultipleJSON([
+          './assets/config/extent_lyon.json',
+          './assets/config/frame3D_planars.json',
+          './assets/config/server/geocoding_server.json',
+          './assets/config/layer/base_maps.json',
+        ])
+        .then((configs) => {
+          // http://proj4js.org/
+          // define a projection as a string and reference it that way
+          // the definition of the projection should be in config TODO_ISSUE
+          udvizBrowser.proj4.default.defs(
+            configs['extent_lyon'].crs,
+            '+proj=lcc +lat_1=45.25 +lat_2=46.75' +
+              ' +lat_0=46 +lon_0=3 +x_0=1700000 +y_0=5200000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs'
+          );
 
-        const extent = new udvizBrowser.itowns.Extent(
-          configs['extent_lyon'].crs,
-          parseInt(configs['extent_lyon'].west),
-          parseInt(configs['extent_lyon'].east),
-          parseInt(configs['extent_lyon'].south),
-          parseInt(configs['extent_lyon'].north)
-        );
+          const extent = new udvizBrowser.itowns.Extent(
+            configs['extent_lyon'].crs,
+            parseInt(configs['extent_lyon'].west),
+            parseInt(configs['extent_lyon'].east),
+            parseInt(configs['extent_lyon'].south),
+            parseInt(configs['extent_lyon'].north)
+          );
 
-        const frame3DPlanar = new udvizBrowser.Frame3DPlanar(
-          extent,
-          configs['frame3D_planars'][2]
-        );
-
-        // /// ADD LAYERS
-
-        udvizBrowser.addBaseMapLayer(
-          configs['base_maps'][0],
-          frame3DPlanar.itownsView,
-          extent
-        );
-
-        udvizBrowser.addLights(frame3DPlanar.getScene());
-        udvizBrowser.initRenderer(
-          frame3DPlanar.getRenderer(),
-          new udvizBrowser.THREE.Color(0x6699cc)
-        );
-
-        // //// REQUEST SERVICE
-        const requestService = new udvizBrowser.RequestService();
-
-        // //// GEOCODING EXTENSION
-        const geocodingView = new udvizBrowser.Widget.Server.GeocodingView(
-          new udvizBrowser.Widget.Server.GeocodingService(
-            requestService,
+          const frame3DPlanar = new udvizBrowser.Frame3DPlanar(
             extent,
-            configs['geocoding_server']
-          ),
-          frame3DPlanar.getItownsView()
-        );
+            configs['frame3D_planars'][2]
+          );
 
-        frame3DPlanar.appendToUI(geocodingView.html());
-      });
+          // /// ADD LAYERS
+
+          udvizBrowser.addBaseMapLayer(
+            configs['base_maps'][0],
+            frame3DPlanar.itownsView,
+            extent
+          );
+
+          // //// REQUEST SERVICE
+          const requestService = new udvizBrowser.RequestService();
+
+          // //// GEOCODING EXTENSION
+          const geocodingView = new udvizBrowser.Widget.Server.GeocodingView(
+            new udvizBrowser.Widget.Server.GeocodingService(
+              requestService,
+              extent,
+              configs['geocoding_server']
+            ),
+            frame3DPlanar.getItownsView()
+          );
+
+          frame3DPlanar.appendToUI(geocodingView.html());
+        });
     </script>
   </body>
 </html>
