+++ packages/browser/src/Component/Widget/Server/Server.js
@@ -2,17 +2,17 @@
 
 // Collection of Itowns Widget using an external server
 
-export { AuthenticationService } from './Authentication/services/AuthenticationService';
+// export { AuthenticationService } from './Authentication/services/AuthenticationService';
 
-export { AuthenticationView } from './Authentication/views/AuthenticationView';
+// export { AuthenticationView } from './Authentication/views/AuthenticationView';
 
 export { GeocodingService } from './Geocoding/services/GeocodingService';
 
 export { GeocodingView } from './Geocoding/views/GeocodingView';
 
-export { SparqlQueryWindow } from './SPARQL/View/SparqlQueryWindow';
+// export { SparqlQueryWindow } from './SPARQL/View/SparqlQueryWindow';
 
-export { SparqlEndpointResponseProvider } from './SPARQL/Service/SparqlEndpointResponseProvider';
+// export { SparqlEndpointResponseProvider } from './SPARQL/Service/SparqlEndpointResponseProvider';
 
-import * as Document from './Document/Document';
-export { Document };
+// import * as Document from './Document/Document';
+// export { Document };
