+++ packages/browser/src/Component/Widget/Widget.js
@@ -1,24 +1,24 @@
-export { CityObjectModule } from './CityObjects/CityObjectModule';
+// export { CityObjectModule } from './CityObjects/CityObjectModule';
 
-export { CityObjectProvider } from './CityObjects/ViewModel/CityObjectProvider';
+// export { CityObjectProvider } from './CityObjects/ViewModel/CityObjectProvider';
 
-export { LayerChoice } from './LayerChoice/views/LayerChoice';
+// export { LayerChoice } from './LayerChoice/views/LayerChoice';
 
-export { CameraPositioner } from './CameraPositioner/CameraPositioner';
+// export { CameraPositioner } from './CameraPositioner/CameraPositioner';
 
-export { TemporalModule } from './Temporal/TemporalModule';
+// export { TemporalModule } from './Temporal/TemporalModule';
 
-export { $3DTemporalBatchTable } from './Temporal/Model/3DTemporalBatchTable.js';
+// export { $3DTemporalBatchTable } from './Temporal/Model/3DTemporalBatchTable.js';
 
-export { $3DTemporalBoundingVolume } from './Temporal/Model/3DTemporalBoundingVolume.js';
+// export { $3DTemporalBoundingVolume } from './Temporal/Model/3DTemporalBoundingVolume.js';
 
-export { $3DTemporalTileset } from './Temporal/Model/3DTemporalTileset.js';
+// export { $3DTemporalTileset } from './Temporal/Model/3DTemporalTileset.js';
 
-export { SlideShow } from './SlideShow/SlideShow';
+// export { SlideShow } from './SlideShow/SlideShow';
 
-export { Debug3DTilesView } from './3DTilesDebug/views/Debug3DTilesView';
+// export { Debug3DTilesView } from './3DTilesDebug/views/Debug3DTilesView';
 
-export { BaseMap } from './BaseMap/BaseMapWindow';
+// export { BaseMap } from './BaseMap/BaseMapWindow';
 
 import * as Server from './Server/Server';
 export { Server };
